Similarly, relocate compiler invocation into a separate file
